// src/app/models/ims/inventory-pack-operation-detail.ts
export class InventoryPackOperationDetail {
    id:number;
    packOperationId: number;
    areaId: number;
    productId: number;
    packId: number;
    pack: string;
    unitsNumber: number;
    unitsShipped: number
    receivedUnits: number;
    indOrigin: boolean;
    qty: number;
    totals: number;
}